Code tidy2 (#236)
* move gyro headers to gyro folder
* split gyro drivers declarations and definitions
* tidy gyro delay usage to internal hal
* tidy i2c/spi bus drivers
* tidy vtx
* format sensors
* rates fix + sensor read sanitization
* rates unittests fix
* fft analyzer tidy

+++ lib/Espfc/src/Device/BusSlave.hpp
@@ -11,7 +11,7 @@ class BusSlave : public BusDevice, public BusAwareDevice
 public:
   BusSlave();
 
-  int begin(BusDevice* dev, int addr);
+  int begin(BusDevice* dev, uint8_t addr);
 
   BusType getType() const override;
